[slow] Block/unblock interface
The user tool can now keep us informed of threads it does not want to
run by simply calling block()/unblock().

Since the tool gets to decide what thread to switch to, this may seem
unnecessary. However, remember that we ask for switches on uncaptures,
so if we do not know which threads are blocked, we'll ask when the tool
has no threads to give to us.