[tests] Add rep test
At some point, spin-fast failed if we added a switchcall before a
rep-prefixed instruction. This is a minimal test that would expose this
bug, using both rep stosb and repnz scasb. I thought Pin might be using
internal context registers to track the state of rep instructions, but
this seems to work fine, all repped instructions are actually
interruptible, and their internal state is in regular registers. So
probably this happened because of a bug that mishandled those registers.