[lib] Reorganize files around a common core
There's no point in having fast and slow be completely separate
implementations. Instead, we can just specialize the tracing and keep
most of the executor handling the same.

This commit breaks fast completely (and disables it for now), with the
goal of splitting tracing and context handling in the slow
implementation.