[spin/slow_tracing] Make setReg() subsume executeAt()
This is a bit more invasive than with fast tracing. Now we always save
the full context before every switchcall, pass and modify our local
copy to switchcalls, and ExecuteAt on that context after the switchcall
if we need to.

In theory, this is slower because it always materializes the context. In
practice, all our ordspecsim switchcalls in materialize the context, so
it's about the same speed.

Beyond paving the way to kill executeAt(), this changeset:
- Removes header differences between slow and fast modes, so no need to
  compile two version of every file that includes spin.h.
- Allows getContext() to always return the right context, even if it's
  called on the current thread (so long as it's called from a
  switchcall).
- As a result, lets us get rid of having to pass ThreadContext* as a
  switchcall argument, which would free up a reg.