[spin] Let syscallEnterCallback disallow uncaptures
**This is an API-breaking change.**

syscallEnterCallback's return value now determines whether the thread
should be uncaptured before taking the syscall, and captured after it
comes back.

The whole reason for the uncapture/capture rigmarole is that some
syscalls are blocking. If we didn't do this and a syscall blocked in the
kernel, the tool would deadlock. So until now spin has done the
uncapture/capture dance for all syscalls.

This change lets the tool determine which syscalls go through the
uncapture/capture cycle instead. If the tool disallows uncaptures for a
syscall, the thread taking the syscall retains the executor role.

FYI, this is my second try at implementing this functionality--in
principle it's possible to implement this completely at the tool by
stalling the simulation while a thread is out for a syscall. However, I
ended up with a complex, brittle mess of racy code. Coopting spin into
this avoids any concurrency to begin with, so the tool-level code is
much simpler.