[fast,scons] Don't depend on AVX, decide whether to save/restore XMMs or YMMs at instrumentation time
The goal of this commit is to have one compiled version of libspin that
can run on machines with AVX as well as older machines that lack AVX.

Even if one tries to compile applications/benchmarks without use of AVX,
if the application runs on a machine with AVX, inevitably the C runtime or
some system library code will access the YMM registers.  So if we want
to run libspin on a machine with AVX, we need to support accesses
from userspace to YMMs.  The only way to have one simulator build that
runs on both AVX and non-AVX machines is to dynamically detect whether
the machine has AVX and change our behavior for what registers to
save/restore accordingly.