[spin] Extend reading/writing XMM/YMM regs to slow_tracing
This should make the behavior of {set,get}Reg for XMM/YMM
registers the same between fast_tracing and slow_tracing.
When using these for floating point computations, these write (or
read) the binary representation of the value (_not_ the floating
point value) to/from XMM/YMM registers. Pintools will need to
convert to/from the binary representation as needed.